# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Goat Nutrition App API
  version: 0.1.0
  description: |
    Minimal OpenAPI for key endpoints. Auth uses cookie session (Flask-Login).
servers:
  - url: http://localhost:5001
paths:
...
  /api/finance/costs:
    get:
      summary: List cost entries
      responses:
        '200':
          description: Cost entry list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceEntryListResponse'
    post:
      summary: Create cost entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinanceEntryPayload'
      responses:
        '201': { description: Created }
        '400': { description: Validation error }
  /api/finance/costs/{entry_id}:
    put:
      summary: Update cost entry
      parameters:
        - in: path
          name: entry_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinanceEntryUpdate'
      responses:
        '200': { description: Updated }
        '404': { description: Not found }
    delete:
      summary: Delete cost entry
      parameters:
        - in: path
          name: entry_id
          required: true
          schema:
            type: integer
      responses:
        '200': { description: Deleted }
        '404': { description: Not found }
  /api/finance/costs/bulk-import:
    post:
      summary: Bulk import cost entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entries:
                  type: array
                  items:
                    $ref: '#/components/schemas/FinanceEntryPayload'
      responses:
        '201': { description: Imported }
        '400': { description: Validation error }
  /api/finance/revenues:
    get:
      summary: List revenue entries
      responses:
        '200':
          description: Revenue entry list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceEntryListResponse'
    post:
      summary: Create revenue entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinanceEntryPayload'
      responses:
        '201': { description: Created }
        '400': { description: Validation error }
  /api/finance/revenues/{entry_id}:
    put:
      summary: Update revenue entry
      parameters:
        - in: path
          name: entry_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinanceEntryUpdate'
      responses:
        '200': { description: Updated }
        '404': { description: Not found }
    delete:
      summary: Delete revenue entry
      parameters:
        - in: path
          name: entry_id
          required: true
          schema:
            type: integer
      responses:
        '200': { description: Deleted }
        '404': { description: Not found }
  /api/finance/revenues/bulk-import:
    post:
      summary: Bulk import revenue entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entries:
                  type: array
                  items:
                    $ref: '#/components/schemas/FinanceEntryPayload'
      responses:
        '201': { description: Imported }
        '400': { description: Validation error }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    FinanceEntryPayload:
      type: object
      required: [recorded_at, category, amount]
      properties:
        recorded_at: { type: string, format: date-time }
        category: { type: string }
        subcategory: { type: string }
        label: { type: string }
        amount: { type: number }
        currency: { type: string }
        sheep_id: { type: integer, nullable: true }
        breed: { type: string, nullable: true }
        age_months: { type: integer, nullable: true }
        lactation_number: { type: integer, nullable: true }
        production_stage: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        extra_metadata:
          type: object
          additionalProperties: true
    FinanceEntryUpdate:
      allOf:
        - $ref: '#/components/schemas/FinanceEntryPayload'
      required: []
    FinanceEntry:
      allOf:
        - $ref: '#/components/schemas/FinanceEntryPayload'
      properties:
        id: { type: integer }
    FinanceEntryListResponse:
      type: object
      properties:
        total: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/FinanceEntry'
...
security:
  - cookieAuth: []